requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
PyPDF2>=3.0.1
fake-useragent>=1.1.1
tldextract>=5.0.1
urllib3>=2.0.7
pyyaml>=6.0.1
click>=8.1.7
